+++ examples/example.cpp
@@ -72,7 +72,7 @@ int main()
                     true,
                     bourne::json({"key", "value"})),
                 "nested objects", {
-                    "is_cool", true
+                    "are_cool", true
                 }
             }
         };
@@ -91,8 +91,8 @@ int main()
 
     // Nested elements can be accessed like so:
     std::cout << std::endl << "accessing nested keys" << std::endl;
-    std::cout << "json[\"key3\"][\"nested objects\"][\"is_cool\"]: "
-              << json["key3"]["nested objects"]["is_cool"].to_bool()
+    std::cout << "json[\"key3\"][\"nested objects\"][\"are_cool\"]: "
+              << json["key3"]["nested objects"]["are_cool"].to_bool()
               << std::endl;
 
     // Elements can be iterated.