+++ src/bourne/detail/error_tags.hpp
@@ -1,0 +1,16 @@
+// Copyright (c) Steinwurf ApS 2016.
+// All Rights Reserved
+//
+// Distributed under the "BSD License". See the accompanying LICENSE.rst file.
+
+BOURNE_ERROR_TAG(unused, "reserved since errors should never be zero :)")
+BOURNE_ERROR_TAG(parse_found_multiple_unstructured_elements, "Found multiple unstructured elements.")
+BOURNE_ERROR_TAG(parse_object_expected_colon, "Expected \":\"")
+BOURNE_ERROR_TAG(parse_object_expected_comma, "Expected \",\"")
+BOURNE_ERROR_TAG(parse_array_expected_comma_or_closing_bracket, "Expected \",\" or \"]\"")
+BOURNE_ERROR_TAG(parse_string_expected_hex_char, "Expected hex character")
+BOURNE_ERROR_TAG(parse_number_expected_number_for_component, "Expected number for component")
+BOURNE_ERROR_TAG(parse_number_unexpected_char, "Unexpected character")
+BOURNE_ERROR_TAG(parse_boolean_expected_true_or_false, "Expected \"true\" or \"false\"")
+BOURNE_ERROR_TAG(parse_null_expected_null, "Expected \"null\"")
+BOURNE_ERROR_TAG(parse_next_unexpected_char, "Unknown starting character.")