+++ NEWS.rst
@@ -6,7 +6,9 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Patch: Fixed multiple unhandled corner cases for parser.
+* Minor: Added custom error codes for parser.
+* Minor: Lowered CMake version requirement.
 
 7.2.0
 -----